requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
gunicorn==20.1.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
phonenumbers==8.13.8
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
zipp==3.15.0
